+++ pydash/pydash_app/user/entity.py
@@ -1,11 +1,14 @@
 from werkzeug.security import generate_password_hash
 from werkzeug.security import check_password_hash
 from .verification_code import VerificationCode
-from pydash_app.user import check_password_requirements
 
 import uuid
 import flask_login
 import persistent
+
+
+_MINIMUM_PASSWORD_LENGTH1 = 8
+_MINIMUM_PASSWORD_LENGTH2 = 12
 
 
 class User(persistent.Persistent, flask_login.UserMixin):
@@ -82,13 +85,26 @@ class User(persistent.Persistent, flask_login.UserMixin):
 
     def set_password(self, password):
 
-        if not check_password_requirements(password):
+        if not self._check_password_requirements(password):
             raise ValueError("Supplied password does not meet requirements")
 
         self.password_hash = generate_password_hash(password)
 
     # Required because `multi_indexed_collection` puts users in a set, that needs to order its keys for fast lookup.
     # Because the IDs are unchanging integer values, use that.
+
+    def _check_password_requirements(self, password):
+        rules1 = [lambda xs: any(x.isupper() for x in xs),
+                  lambda xs: any(not x.isalpha() for x in xs),
+                  lambda xs: len(xs) >= _MINIMUM_PASSWORD_LENGTH1
+                  ]
+        rules2 = [lambda xs: len(xs) >= _MINIMUM_PASSWORD_LENGTH2]
+        alternatives = [rules1, rules2]
+
+        def func(rules):
+            return all(rule(password) for rule in rules)
+
+        return any(func(alternative) for alternative in alternatives)
 
     def __lt__(self, other):
         """
